requirements.txt
roboflow==1.2.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.1
pillow==10.0.1
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
pytesseract==0.3.10 